refactor : NPE 방지와 싱글톤 Validator를 이용한 검증 책임 위임
